+++ nilearn/_utils/__init__.py
@@ -1,7 +1,7 @@
 
 from .niimg_conversions import check_niimg, concat_niimgs, check_niimgs
 
-from .niimage import new_img, load_img, is_img, _get_shape, _repr_niimgs, \
+from .niimg import new_img, load_img, is_img, _get_shape, _repr_niimgs, \
         copy_img
 
 from .numpy_conversions import as_ndarray
