+++ nilearn/_utils/niimg.py
